refactor: seperate snapshot provider
