+++ src/perses-dashboard/react/SnapshotViewDashboard.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { Box } from '@mui/material'
+import { Box, GlobalStyles } from '@mui/material'
 import {
   DashboardProvider,
   DatasourceStoreProvider,
@@ -8,11 +8,56 @@ import {
   type ViewDashboardProps,
 } from '@perses-dev/dashboards'
 import { DashboardApp } from '@perses-dev/dashboards/dist/views/ViewDashboard/DashboardApp'
-import { DEFAULT_REFRESH_INTERVAL } from '@perses-dev/core'
+import { DEFAULT_REFRESH_INTERVAL, type DashboardResource } from '@perses-dev/core'
 import { ErrorAlert, ErrorBoundary, combineSx } from '@perses-dev/components'
 import { TimeRangeProviderBasic, usePluginBuiltinVariableDefinitions } from '@perses-dev/plugin-system'
 import { getSnapshotEmbed } from '../snapshot/isSnapshotDashboard'
-import { snapshotTimeRangeToAbsoluteValue } from '../snapshot/resolveSnapshotTimeRange'
+import { normalizeSnapshotTimeRange, snapshotTimeRangeToAbsoluteValue } from '../snapshot/resolveSnapshotTimeRange'
+import { SnapshotProvider } from './SnapshotContext'
+
+export function getSnapshotGlobalStyles(): Record<string, Record<string, unknown>> {
+  return {
+    'body.dashboard-snapshot-mode [data-testid="dashboard-toolbar"] button[aria-label="Refresh"]': {
+      display: 'none',
+    },
+    'body.dashboard-snapshot-mode [data-testid="dashboard-toolbar"] > .MuiBox-root:nth-of-type(2) > .MuiStack-root > .MuiStack-root > :first-child':
+      {
+        pointerEvents: 'none',
+        opacity: 0.85,
+      },
+  }
+}
+
+export function getSnapshotViewRemountKey(dashboard: DashboardResource, saveRefreshToken: number): string {
+  const snapshotEmbed = getSnapshotEmbed(dashboard)
+  const snapshotFrozenRange = normalizeSnapshotTimeRange(snapshotEmbed?.timeRange)
+  const snapshotTimeRangeKey = snapshotFrozenRange ? `${snapshotFrozenRange.from}-${snapshotFrozenRange.to}` : ''
+  return `${dashboard.metadata.name}-${saveRefreshToken}-${snapshotTimeRangeKey}`
+}
+
+/** Snapshot-only shell: context, body class, and toolbar overrides. */
+export function SnapshotDashboardShell({
+  dashboard,
+  children,
+}: {
+  dashboard: DashboardResource
+  children: React.ReactNode
+}): JSX.Element {
+  React.useEffect(() => {
+    const { body } = document
+    body.classList.add('dashboard-snapshot-mode')
+    return () => {
+      body.classList.remove('dashboard-snapshot-mode')
+    }
+  }, [])
+
+  return (
+    <SnapshotProvider dashboard={dashboard}>
+      <GlobalStyles styles={getSnapshotGlobalStyles()} />
+      {children}
+    </SnapshotProvider>
+  )
+}
 
 /**
  * Snapshot dashboard view: time range comes only from `spec.snapshot.timeRange`,
@@ -74,7 +119,7 @@ export default function SnapshotViewDashboard(props: ViewDashboardProps): JSX.El
       },
     ]
     if (data) {
-      data.forEach((def) => result.push(def))
+      data.forEach((def) => result.push(def as (typeof result)[number]))
     }
     return result
   }, [dashboardResource.metadata.name, dashboardResource.metadata.project, data])
